move matrixvt_det to models
